refactor: delete modules/requests/pages/routes
