+++ web-client/src/modules/requests/containers/FindRequestsContainer.tsx
@@ -32,7 +32,7 @@ import { COLORS } from 'src/theme/colors';
 import styled from 'styled-components';
 
 import RequestItem from '../components/RequestItem';
-import { OpenRequestsLocation } from '../pages/routes/OpenRequestsRoute/constants';
+import { OpenRequestsLocation } from '../constants';
 
 const { TabPane } = Tabs;
