--- web-client/src/modules/requests/pages/routes/AcceptedRequestsRoute/AcceptedRequestsRoute.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-
-import AcceptedRequestsContainer from '../../../containers/AcceptedRequestsContainer';
-
-const AcceptedRequestsRoute: React.FC = (): React.ReactElement => (
-  <AcceptedRequestsContainer />
-);
-export default AcceptedRequestsRoute;
