--- web-client/src/modules/requests/pages/routes/AcceptedRequestsRoute/constants.tsx
@@ -1,3 +1,0 @@
-import Location from 'react-app-location';
-
-export const AcceptedRequestsLocation = new Location('/requests/accepted');
