--- web-client/src/modules/requests/pages/routes/FindRequestsRoute/FindRequestsRoute.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import FindRequestsContainer from '../../../containers/FindRequestsContainer';
-
-const FindRequestsRoute: React.FC = () => <FindRequestsContainer />;
-
-export default FindRequestsRoute;
