--- web-client/src/modules/requests/pages/routes/NewRequestsRoute/NewRequestsRoute.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import NewRequestsContainer from '../../../containers/NewRequestsContainer';
-
-const NewRequestsRoute: React.FC = () => <NewRequestsContainer />;
-
-export default NewRequestsRoute;
