--- web-client/src/modules/requests/pages/routes/NewRequestsRoute/constants.tsx
@@ -1,3 +1,0 @@
-import Location from 'react-app-location';
-
-export const NewRequestsLocation = new Location('/requests/new');
