--- web-client/src/modules/requests/pages/routes/PostsRoute/constants.tsx
@@ -1,6 +1,0 @@
-import Location from 'react-app-location';
-
-/*
- * :type is either 'ask' or 'offer'
- */
-export const MainLocation = new Location('/requests/main/:type');
